pyproject.toml
[project]
name = "simple-code-cli"
version = "0.1.0"
description = "A tiny Claude Code-inspired terminal coding agent using Vertex AI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.scripts]
simple-code = "simple_code_cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
